Implement "implicit transactions" as used by redis-rb
When using Redis::Cluster#watch from redis-rb, it does not actually pass
the user block to RedisClient::Cluster. Instead, it actually manages the
process of WATCH/UNWATCH itself, and issues call_v(['WATCH', ...]) and
call_v(['UNWATCH']).

To support this, we need to detect when WATCH is called and begin an
"implicit transaction", within which we pin to a slot (and actually pin
to a connection as well, in the case of the Pooled backend). Until
UNWATCH/EXEC is issued, the connection cannot be used for access to
other keys not on the transaction node.

That's implemented by having the client keep a @transaction member, and
set it when a WATCH command is issued.