Allow node topology instances to process _updates_ to topology
At the moment, we throw away the whole ::Node structure every time we
call CLUSTER NODES. This has a couple of unfortunate side effects:

* If one node has failed and been replaced, we wind up disconnecting
  from _all_ of the nodes, even if they're perfectly healthy. This can
  cause an unexpected connection storm on the servers.
* Because of this, we also replace the underlying RedisClient instances,
  which are responsible for maintaining circuit breakers. We actually
  therefore wind up constructing a new circuit breaker every time we
  process any topology change, and so the circuit breaker will never
  open. This, again, can lead to undesirable load on the Redis servers.

This patch is the first step towards correcting the problem. It makes
the Topology class instaces have a method `process_topology_update!`
which gives them the ability to maintain RedisClient instances that
don't need to be changed. However, we don't actually yet call this
outside of #initialize.